Revert "refactor: 운동거리 int로 변경"
This reverts commit a3830e891d12d01eadcdf8c1d942a46fc24b4f80.

+++ src/main/java/clofi/runningplanet/running/dto/RecordFindAllResponse.java
@@ -4,7 +4,7 @@ import clofi.runningplanet.running.domain.Record;
 
 public record RecordFindAllResponse(
 	Long id,
-	int runDistance,
+	double runDistance,
 	int day
 ) {
 	public RecordFindAllResponse(Record record) {
